reorganize the rpc stuff
